[Feature/Solid] Explain Single Responsibility Principle (SRP)
- Created a `ReadMe.md` file explaining the Single Responsibility Principle (SRP) and its importance.
- Added `Before-SRP.swift` demonstrating code with SRP violation where the `Journal` class handles both
  journal management and persistence responsibilities.
- Added `After-SRP.swift` refactoring the code to adhere to SRP by separating the persistence logic into
  a dedicated `Persistence` class, allowing the `Journal` class to focus solely on managing journal entries.

This commit illustrates the benefits of SRP by providing a clear before-and-after comparison.
